[Core] [Source] [Metadata] Refactor metadata info
