+++ core/datacap-ui/src/views/layouts/metadata/MetadataContainer.vue
@@ -3,10 +3,13 @@
     <div class="hidden space-y-6 w-full md:block">
       <div class="flex flex-col space-y-8 lg:flex-row lg:space-x-6 lg:space-y-0">
         <aside class="-mx-4 w-[200px]">
-          <MetadataSidebar :code="code"/>
+          <MetadataSidebar :code="code" @change="handlerChange"/>
         </aside>
         <div class="flex-1">
-          <RouterView/>
+          <Card v-if="!dataInfo" :body-class="'p-8'" :hidden-title="true">
+            <Alert :description="$t('source.tip.notSelectedNode')"/>
+          </Card>
+          <MetadataContent v-else/>
         </div>
       </div>
     </div>
@@ -16,24 +19,37 @@
 <script lang="ts">
 import { defineComponent } from 'vue'
 import MetadataSidebar from '@/views/layouts/metadata/components/MetadataSidebar.vue'
-import { useRouter } from 'vue-router'
-import DatabaseService from '@/services/database.ts'
+import MetadataContent from '@/views/layouts/metadata/components/MetadataContent.vue'
 import { StructureModel } from '@/model/structure.ts'
+import router from '@/router'
+import Card from '@/views/ui/card'
+import Alert from '@/views/ui/alert'
 
 export default defineComponent({
   name: 'MetadataContainer',
   components: {
-    MetadataSidebar
+    MetadataContent,
+    MetadataSidebar,
+    Card,
+    Alert
   },
   data()
   {
     return {
-      code: null as string | null
+      code: null as string | null,
+      dataInfo: null as StructureModel | null
     }
   },
   created()
   {
-    this.code = this.$route.params?.code as string
+    this.code = this.$route.params?.source as string
+  },
+  methods: {
+    handlerChange(node: StructureModel)
+    {
+      this.dataInfo = node
+      router.push(`/admin/source/manager/${ this.code }/info/${ node.code }`)
+    }
   }
 })
 </script>
